Add graphic design files for the new banknotes
Also remove the bundles as unused

Adds the following files used in development of the new banknotes:
* A GIMP palette
* A markdown styleguide
* An XCF banknote template

Full XCF of all banknotes is being kept out for now to reduce the size
of the repository. The new banknotes are all still licensed under
CC-BY-SA_4.0 as the original MG notes were, they're just my work instead
of the original author(s').